Rusty Skylines Alpha v1.5.18b Feb 15th 22:22/23/24
Finished the Motion Vector Texture completely. [major]
Added TAA into the RT Shadows Upscaling pass, very light, no unreal engine garbage, I promise you. [major]

Added Rayon to my game and parallelized the AI cars driving around mindlessly every tick! [minor]
Runs at 10x speed with 16k Cars, though the CPU RENDER performance was very bad then, because it needs to prepare all that data for the GPU and upload it every frame.
And for 16 thousand cars, no frustum culling, unoptimized in general and not parallel, it starts to struggle ofc.

Stabilized the "achieved_time_speed" against hysteresis, as it often jumped between 0.6 and 1.6. [minor]
That is because fps is 100 and tps is 60, so some frames it is not running a tick, and some frames it is!
So it just chooses to wobble and jitter its brains out. (ew)

Fixed an overflow bug in dist2 calculation, happening when you somehow teleport tons of chunks away, causing the multiplication to overflow! [minor]

***
What the hell is the time?? 22:33? I wanted to write it at 22:22. The last days, I became more and more sleep-deprived and noticed that I LOOK MORE GERMAN WTF
Germans are sleep-deprived, aight? But I don't want to be sleepy all day 😡 22:34 See? I am going to sleep NOW instead of in 1.5 hours when I just do some bullshit anyway.

+++ src/helpers/positions.rs
@@ -43,8 +43,8 @@ impl ChunkCoord {
     }
     #[inline]
     pub fn dist2(&self, other: &ChunkCoord) -> u32 {
-        let dx = self.x - other.x;
-        let dz = self.z - other.z;
+        let dx = self.x as i64 - other.x as i64;
+        let dz = self.z as i64 - other.z as i64;
         (dx * dx + dz * dz) as u32
     }
 }
